+++ tools/installer/imagegen/diskutils/diskutils.go
@@ -8,40 +8,48 @@ package diskutils
 import (
 	"encoding/json"
 	"fmt"
-	"os"
-	"path/filepath"
 	"regexp"
 	"strconv"
 	"strings"
-	"time"
 
-	"installer/imagegen/configuration"
 	"installer/internal/file"
-	"installer/internal/logger"
-	"installer/internal/retry"
 	"installer/internal/shell"
+)
 
-	"github.com/sirupsen/logrus"
+// Unit to byte conversion values
+const (
+	B  = 1
+	KB = 1000
+	MB = 1000 * 1000
+	GB = 1000 * 1000 * 1000
+	TB = 1000 * 1000 * 1000 * 1000
+
+	KiB = 1024
+	MiB = 1024 * 1024
+	GiB = 1024 * 1024 * 1024
+	TiB = 1024 * 1024 * 1024 * 1024
+)
+
+// Boot type constants
+const (
+	EFIPartitionType    = "efi"
+	LegacyPartitionType = "legacy"
 )
 
 var (
-	// When calling mkfs, the default options change depending on the host OS you are running on and typically match
-	// what the distro has decided is best for their OS. For example, for ext2/3/4, the defaults are stored in
-	// /etc/mke2fs.conf.
-	// However, when building Azure Linux images, the defaults should be as consistent as possible and should only contain
-	// features that are supported on Azure Linux.
-	DefaultMkfsOptions = map[string][]string{
-		"ext2": {"-b", "4096", "-O", "none,sparse_super,large_file,filetype,resize_inode,dir_index,ext_attr"},
-		"ext3": {"-b", "4096", "-O", "none,sparse_super,large_file,filetype,resize_inode,dir_index,ext_attr,has_journal"},
-		// grub2 doesn't recognize ext4 with metadata_csum_seed enabled
-		// ^metadata_csum_seed disables filesystem to store the metadata checksum seed in the superblock, hence disables changing uuid of mounted filesystem
-		"ext4": {"-b", "4096", "-O", "none,sparse_super,large_file,filetype,resize_inode,dir_index,ext_attr,has_journal,extent,huge_file,flex_bg,metadata_csum,64bit,dir_nlink,extra_isize,^metadata_csum_seed"},
+	sizeAndUnitRegexp = regexp.MustCompile(`(\d+)((Ki?|Mi?|Gi?|Ti?)?B)`)
+
+	unitToBytes = map[string]uint64{
+		"B":   B,
+		"KB":  KB,
+		"MB":  MB,
+		"GB":  GB,
+		"TB":  TB,
+		"KiB": KiB,
+		"MiB": MiB,
+		"GiB": GiB,
+		"TiB": TiB,
 	}
-
-	partedVersionRegex = regexp.MustCompile(`^parted \(GNU parted\) (\d+)\.(\d+)`)
-
-	// The default partition name used when the version of `parted` is too old (<3.5).
-	LegacyDefaultParitionName = "primary"
 )
 
 type blockDevicesOutput struct {
@@ -62,84 +70,20 @@ type SystemBlockDevice struct {
 	Model       string // Example: Virtual Disk
 }
 
-type partitionInfoOutput struct {
-	Devices []PartitionInfo `json:"blockdevices"`
-}
+// SystemBootType returns the current boot type of the system being ran on.
+func SystemBootType() (bootType string) {
+	// If a system booted with EFI, /sys/firmware/efi will exist
+	const efiFirmwarePath = "/sys/firmware/efi"
 
-type PartitionInfo struct {
-	Name              string `json:"name"`       // Example: nbd0p1
-	Path              string `json:"path"`       // Example: /dev/nbd0p1
-	PartitionTypeUuid string `json:"parttype"`   // Example: c12a7328-f81f-11d2-ba4b-00a0c93ec93b
-	FileSystemType    string `json:"fstype"`     // Example: vfat
-	Uuid              string `json:"uuid"`       // Example: 4BD9-3A78
-	PartUuid          string `json:"partuuid"`   // Example: 7b1367a6-5845-43f2-99b1-a742d873f590
-	Mountpoint        string `json:"mountpoint"` // Example: /mnt/os/boot
-	PartLabel         string `json:"partlabel"`  // Example: boot
-	Type              string `json:"type"`       // Example: part
-}
-
-type loopbackListOutput struct {
-	Devices []loopbackDevice `json:"loopdevices"`
-}
-
-type loopbackDevice struct {
-	Name        string `json:"name"`
-	BackingFile string `json:"back-file"`
-}
-
-const (
-	// AutoEndSize is used as the disk's "End" value to indicate it should be picked automatically
-	AutoEndSize = 0
-
-	EfiSystemPartitionTypeUuid = "c12a7328-f81f-11d2-ba4b-00a0c93ec93b"
-	BiosBootPartitionTypeUuid  = "21686148-6449-6e6f-744e-656564454649"
-)
-
-const (
-	// mappingFilePath is used for device mapping paths
-	mappingFilePath = "/dev/mapper/"
-
-	// maxPrimaryPartitionsForMBR is the maximum number of primary partitions
-	// allowed in the case of MBR partition
-	maxPrimaryPartitionsForMBR = 4
-
-	// name of all possible partition types
-	primaryPartitionType  = "primary"
-	extendedPartitionType = "extended"
-	logicalPartitionType  = "logical"
-)
-
-// Unit to byte conversion values
-// See https://www.gnu.org/software/parted/manual/parted.html#unit
-const (
-	B  = 1
-	KB = 1000
-	MB = 1000 * 1000
-	GB = 1000 * 1000 * 1000
-	TB = 1000 * 1000 * 1000 * 1000
-
-	KiB = 1024
-	MiB = 1024 * 1024
-	GiB = 1024 * 1024 * 1024
-	TiB = 1024 * 1024 * 1024 * 1024
-)
-
-var (
-	sizeAndUnitRegexp = regexp.MustCompile(`(\d+)((Ki?|Mi?|Gi?|Ti?)?B)`)
-	diskDevPathRegexp = regexp.MustCompile(`^/dev/(\w+)$`)
-
-	unitToBytes = map[string]uint64{
-		"B":   B,
-		"KB":  KB,
-		"MB":  MB,
-		"GB":  GB,
-		"TB":  TB,
-		"KiB": KiB,
-		"MiB": MiB,
-		"GiB": GiB,
-		"TiB": TiB,
+	exist, _ := file.DirExists(efiFirmwarePath)
+	if exist {
+		bootType = EFIPartitionType
+	} else {
+		bootType = LegacyPartitionType
 	}
-)
+
+	return
+}
 
 // BytesToSizeAndUnit takes a number of bytes and returns friendly representation of a size (for example 100GB).
 func BytesToSizeAndUnit(bytes uint64) string {
@@ -200,622 +144,7 @@ func SizeAndUnitToBytes(sizeAndUnit string) (bytes uint64, err error) {
 	if unitBytes, ok := unitToBytes[unit]; ok {
 		bytes = size * unitBytes
 	} else {
-		err = fmt.Errorf("unknown unit type (%s)", unit)
-		return
-	}
-
-	return
-}
-
-// CreateEmptyDisk creates an empty raw disk in the given working directory as described in disk configuration
-func CreateEmptyDisk(workDirPath, diskName string, maxSize uint64) (diskFilePath string, err error) {
-	diskFilePath = filepath.Join(workDirPath, diskName)
-
-	err = CreateSparseDisk(diskFilePath, maxSize, 0o644)
-	return
-}
-
-// CreateSparseDisk creates an empty sparse disk file.
-func CreateSparseDisk(diskPath string, size uint64, perm os.FileMode) (err error) {
-	// Open and truncate the file.
-	file, err := os.OpenFile(diskPath, os.O_CREATE|os.O_TRUNC|os.O_WRONLY, perm)
-	if err != nil {
-		return fmt.Errorf("failed to create empty disk file:\n%w", err)
-	}
-
-	// Resize the file to the desired size.
-	err = file.Truncate(int64(size * MiB))
-	if err != nil {
-		return fmt.Errorf("failed to set empty disk file's size:\n%w", err)
-	}
-	return
-}
-
-// SetupLoopbackDevice creates a /dev/loop device for the given disk file
-func SetupLoopbackDevice(diskFilePath string) (devicePath string, err error) {
-	logger.Log.Debugf("Attaching Loopback: %v", diskFilePath)
-	stdout, stderr, err := shell.Execute("losetup", "--show", "-f", "-P", diskFilePath)
-	if err != nil {
-		err = fmt.Errorf("failed to create loopback device using losetup:\n%v\n%w", stderr, err)
-		return
-	}
-	devicePath = strings.TrimSpace(stdout)
-	logger.Log.Debugf("Created loopback device at device path: %v", devicePath)
-	return
-}
-
-// BlockOnDiskIO waits until all outstanding operations against a disk complete.
-func BlockOnDiskIO(diskDevPath string) (err error) {
-	maj, min, err := GetDiskIds(diskDevPath)
-	if err != nil {
-		return
-	}
-
-	return BlockOnDiskIOByIds(diskDevPath, maj, min)
-}
-
-func GetDiskIds(diskDevPath string) (maj string, min string, err error) {
-	rawDiskOutput, stderr, err := shell.Execute("lsblk", "--nodeps", "--json", "--output", "NAME,MAJ:MIN", diskDevPath)
-	if err != nil {
-		err = fmt.Errorf("failed to find IDs for disk (%s):\n%v\n%w", diskDevPath, stderr, err)
-		return
-	}
-
-	var blockDevices blockDevicesOutput
-	if rawDiskOutput != "" {
-		err = json.Unmarshal([]byte(rawDiskOutput), &blockDevices)
-		if err != nil {
-			return
-		}
-	}
-
-	if len(blockDevices.Devices) != 1 {
-		err = fmt.Errorf("couldn't find disk IDs for %s (%s), expecting only one result", diskDevPath, rawDiskOutput)
-		return
-	}
-	// MAJ:MIN is returned in the form "1:2"
-	diskIDs := strings.Split(blockDevices.Devices[0].MajMin, ":")
-	if len(diskIDs) != 2 {
-		err = fmt.Errorf("couldn't find disk IDs for %s (%s), couldn't parse MAJ:MIN", diskDevPath, rawDiskOutput)
-		return
-	}
-	maj = diskIDs[0]
-	min = diskIDs[1]
-	return
-}
-
-// BlockOnDiskIOById waits until all outstanding operations against a disk complete.
-func BlockOnDiskIOByIds(debugName string, maj string, min string) (err error) {
-	const (
-		// Indices for values in /proc/diskstats
-		majIdx            = 0
-		minIdx            = 1
-		outstandingOpsIdx = 11
-	)
-
-	logger.Log.Debugf("Flushing all IO to disk")
-	_, _, err = shell.Execute("sync")
-	if err != nil {
-		return
-	}
-
-	logger.Log.Tracef("Searching /proc/diskstats for %s (%s:%s)", debugName, maj, min)
-	for {
-		var (
-			foundEntry     = false
-			outstandingOps = ""
-		)
-
-		// Find the entry with Major#, Minor#, ..., IOs which matches our disk
-		onStdout := func(line string) {
-			// Bail early if we already found the entry
-			if foundEntry {
-				return
-			}
-
-			deviceStatsFields := strings.Fields(line)
-			if maj == deviceStatsFields[majIdx] && min == deviceStatsFields[minIdx] {
-				outstandingOps = deviceStatsFields[outstandingOpsIdx]
-				foundEntry = true
-			}
-		}
-
-		err = shell.NewExecBuilder("cat", "/proc/diskstats").
-			StdoutCallback(onStdout).
-			WarnLogLines(shell.DefaultWarnLogLines).
-			LogLevel(logrus.TraceLevel, logrus.ErrorLevel).
-			Execute()
-		if err != nil {
-			return
-		}
-		if !foundEntry {
-			return fmt.Errorf("couldn't find entry for '%s' in /proc/diskstats", debugName)
-		}
-		logger.Log.Debugf("Outstanding operations on '%s': %s", debugName, outstandingOps)
-
-		if outstandingOps == "0" {
-			break
-		}
-
-		// Sleep breifly
-		time.Sleep(time.Second / 4)
-	}
-	return
-}
-
-// DetachLoopbackDevice detaches the specified disk
-func DetachLoopbackDevice(diskDevPath string) (err error) {
-	logger.Log.Debugf("Detaching Loopback Device Path: %v", diskDevPath)
-	_, stderr, err := shell.Execute("losetup", "-d", diskDevPath)
-	if err != nil {
-		logger.Log.Warnf("Failed to detach loopback device using losetup: %v", stderr)
-	}
-	return
-}
-
-func WaitForLoopbackToDetach(devicePath string, diskPath string) error {
-	if !filepath.IsAbs(diskPath) {
-		return fmt.Errorf("internal error: loopback disk path must be absolute (%s)", diskPath)
-	}
-
-	delay := 120 * time.Millisecond
-	attempts := 10
-	for failures := 0; failures < attempts; failures++ {
-		stdout, _, err := shell.Execute("losetup", "--list", "--json", "--output", "NAME,BACK-FILE")
-		if err != nil {
-			return fmt.Errorf("failed to read loopback list:\n%w", err)
-		}
-
-		var output loopbackListOutput
-		if stdout != "" {
-			err = json.Unmarshal([]byte(stdout), &output)
-			if err != nil {
-				return fmt.Errorf("failed to parse loopback devices list JSON:\n%w", err)
-			}
-		}
-
-		found := false
-		for _, device := range output.Devices {
-			if device.Name == devicePath && device.BackingFile == diskPath {
-				found = true
-				break
-			}
-		}
-
-		if !found {
-			return nil
-		}
-
-		time.Sleep(delay)
-		delay *= 2
-	}
-
-	return fmt.Errorf("timed out waiting for loopback device (%s) for disk (%s) to close", devicePath, diskPath)
-}
-
-// WaitForDevicesToSettle waits for all udev events to be processed on the system.
-// This can be used to wait for partitions to be discovered after mounting a disk.
-func WaitForDevicesToSettle() error {
-	logger.Log.Debugf("Waiting for devices to settle")
-	_, _, err := shell.Execute("udevadm", "settle")
-	if err != nil {
-		return fmt.Errorf("failed to wait for devices to settle:\n%w", err)
-	}
-	return nil
-}
-
-// CreatePartitions creates partitions on the specified disk according to the disk config
-func CreatePartitions(diskDevPath string, disk configuration.Disk, rootEncryption configuration.RootEncryption,
-	diskKnownToBeEmpty bool,
-) (partDevPathMap map[string]string, partIDToFsTypeMap map[string]string, encryptedRoot EncryptedRootDevice, err error) {
-	const timeoutInSeconds = "5"
-	partDevPathMap = make(map[string]string)
-	partIDToFsTypeMap = make(map[string]string)
-
-	// Clear any old partition table info to prevent errors during partition creation
-	if !diskKnownToBeEmpty {
-		_, stderr, err := shell.Execute("sfdisk", "--delete", diskDevPath)
-		if err != nil {
-			logger.Log.Warnf("Failed to clear partition table. Expected if the disk is blank: %v", stderr)
-		}
-	}
-
-	// Create new partition table
-	partitionTableType := disk.PartitionTableType
-	logger.Log.Debugf("Converting partition table type (%v) to parted argument", partitionTableType)
-	partedArgument, err := partitionTableType.ConvertToPartedArgument()
-	if err != nil {
-		err = fmt.Errorf("failed to convert partition table type (%v) to parted argument:\n%w", partitionTableType, err)
-		return
-	}
-	_, stderr, err := shell.Execute("flock", "--timeout", timeoutInSeconds, diskDevPath, "parted", diskDevPath, "--script", "mklabel", partedArgument)
-	if err != nil {
-		err = fmt.Errorf("failed to set partition table type using parted:\n%v\n%w", stderr, err)
-		return
-	}
-
-	usingExtendedPartition := (len(disk.Partitions) > maxPrimaryPartitionsForMBR) && (partitionTableType == configuration.PartitionTableTypeMbr)
-
-	partedSupportsEmptyStringArgs, err := PartedSupportsEmptyString()
-	if err != nil {
-		return
-	}
-
-	// Partitions assumed to be defined in sorted order
-	for idx, partition := range disk.Partitions {
-		partType, partitionNumber := obtainPartitionDetail(idx, usingExtendedPartition)
-		// Insert an extended partition
-		if partType == extendedPartitionType {
-			err = createExtendedPartition(diskDevPath, partitionTableType, disk.Partitions, partIDToFsTypeMap,
-				partDevPathMap, partedSupportsEmptyStringArgs)
-			if err != nil {
-				return
-			}
-
-			// Update partType and partitionNumber
-			partType = logicalPartitionType
-			partitionNumber = partitionNumber + 1
-		}
-
-		partDevPath, err := createSinglePartition(diskDevPath, partitionNumber, partitionTableType, partition, partType,
-			partedSupportsEmptyStringArgs)
-		if err != nil {
-			err = fmt.Errorf("failed to create single partition:\n%w", err)
-			return partDevPathMap, partIDToFsTypeMap, encryptedRoot, err
-		}
-
-		partFsType, err := FormatSinglePartition(partDevPath, partition)
-		if err != nil {
-			err = fmt.Errorf("failed to format partition:\n%w", err)
-			return partDevPathMap, partIDToFsTypeMap, encryptedRoot, err
-		}
-
-		if rootEncryption.Enable && partition.HasFlag(configuration.PartitionFlagDeviceMapperRoot) {
-			encryptedRoot, err = encryptRootPartition(partDevPath, partition, rootEncryption)
-			if err != nil {
-				err = fmt.Errorf("failed to initialize encrypted root:\n%w", err)
-				return partDevPathMap, partIDToFsTypeMap, encryptedRoot, err
-			}
-			partDevPathMap[partition.ID] = GetEncryptedRootVolMapping()
-		} else {
-			partDevPathMap[partition.ID] = partDevPath
-		}
-
-		partIDToFsTypeMap[partition.ID] = partFsType
-	}
-	return
-}
-
-// createSinglePartition creates a single partition based on the partition config
-func createSinglePartition(diskDevPath string, partitionNumber int, partitionTableType configuration.PartitionTableType,
-	partition configuration.Partition, partType string, partedSupportsEmptyStringArgs bool,
-) (partDevPath string, err error) {
-	const (
-		fillToEndOption  = "100%"
-		sFmt             = "%ds"
-		timeoutInSeconds = "5"
-	)
-
-	logicalSectorSize, physicalSectorSize, err := GetSectorSize(diskDevPath)
-	if err != nil {
-		return
-	}
-
-	start := partition.Start * MiB / logicalSectorSize
-	end := partition.End*MiB/logicalSectorSize - 1
-	if partition.End == 0 {
-		end = 0
-	}
-
-	if partType == logicalPartitionType {
-		start = start + 1
-		if end != 0 {
-			end = end + 1
-		}
-	}
-
-	// Check whether the start sector is 4K-aligned
-	start = alignSectorAddress(start, logicalSectorSize, physicalSectorSize)
-
-	logger.Log.Debugf("Input partition start: %d, aligned start sector: %d", partition.Start, start)
-	logger.Log.Debugf("Input partition end: %d, end sector: %d", partition.End, end)
-
-	mkpartArgs := []string{"--timeout", timeoutInSeconds, diskDevPath, "parted", diskDevPath, "--script", "mkpart"}
-
-	switch partitionTableType {
-	case configuration.PartitionTableTypeMbr:
-		// Part type.
-		mkpartArgs = append(mkpartArgs, partType)
-
-	case configuration.PartitionTableTypeGpt:
-		// Partition label.
-		if partition.Name == "" {
-			if partedSupportsEmptyStringArgs {
-				// For parted, you have to specify "" to represent an empty string.
-				mkpartArgs = append(mkpartArgs, `""`)
-			} else {
-				// This version of parted has no way to specify an empty partition name. :-(
-				// So, use the legacy label of "primary" (which was used in Azure Linux 2.0) instead.
-				logger.Log.Warnf("parted version <3.5 does not support empty partition names: using partition name '%s' instead",
-					LegacyDefaultParitionName)
-				mkpartArgs = append(mkpartArgs, LegacyDefaultParitionName)
-			}
-		} else {
-			mkpartArgs = append(mkpartArgs, partition.Name)
-		}
-	}
-
-	fsType := partition.FsType
-	if fsType == "vfat" {
-		// 'parted mkpart' requires value of either 'fat16' or 'fat32'.
-		fsType = "fat32"
-	}
-
-	if fsType != "" {
-		mkpartArgs = append(mkpartArgs, fsType)
-	}
-
-	mkpartArgs = append(mkpartArgs, fmt.Sprintf(sFmt, start))
-
-	if end == 0 {
-		mkpartArgs = append(mkpartArgs, fillToEndOption)
-	} else {
-		mkpartArgs = append(mkpartArgs, fmt.Sprintf(sFmt, end))
-	}
-
-	_, stderr, err := shell.Execute("flock", mkpartArgs...)
-	if err != nil {
-		err = fmt.Errorf("failed to create partition using parted:\n%v\n%w", stderr, err)
-		return "", err
-	}
-
-	// Update kernel partition table information
-	//
-	// There can be a timing issue where partition creation finishes but the
-	// devtmpfs files are not populated in time for partition initialization.
-	// So to deal with this, we call partprobe here to query and flush the
-	// partition table information, which should enforce that the devtmpfs
-	// files are created when partprobe returns control.
-	//
-	// Added flock because "partprobe -s" apparently doesn't always block.
-	// flock is part of the util-linux package and helps to synchronize access
-	// with other cooperating processes. The important part is it will block
-	// if the fd is busy, and then execute the command. Adding a timeout
-	// to prevent us from possibly waiting forever.
-	stdout, stderr, err := shell.Execute("flock", "--timeout", timeoutInSeconds, diskDevPath, "partprobe", "-s", diskDevPath)
-	if err != nil {
-		err = fmt.Errorf("failed to execute partprobe:\n%v\n%w", stderr, err)
-		return "", err
-	}
-	logger.Log.Debugf("Partprobe -s returned: %s", stdout)
-	return InitializeSinglePartition(diskDevPath, partitionNumber, partitionTableType, partition)
-}
-
-// Returns true if the version of 'parted' supports the 'type' session command.
-// Since v3.6
-func PartedSupportsTypeCommand() (bool, error) {
-	major, minor, err := getPartedVersion()
-	if err != nil {
-		return false, err
-	}
-
-	supports := major >= 4 || (major == 3 && minor >= 6)
-	return supports, nil
-}
-
-// Returns if the version of 'parted' supports empty (quoted) string parameters.
-// Specifically, parted v3.5+.
-func PartedSupportsEmptyString() (bool, error) {
-	major, minor, err := getPartedVersion()
-	if err != nil {
-		return false, err
-	}
-
-	supports := major >= 4 || (major == 3 && minor >= 5)
-	return supports, nil
-}
-
-func getPartedVersion() (int, int, error) {
-	stdout, _, err := shell.Execute("parted", "--version")
-	if err != nil {
-		err = fmt.Errorf("failed to get 'parted' version:\n%w", err)
-		return 0, 0, err
-	}
-
-	matches := partedVersionRegex.FindStringSubmatch(stdout)
-	if matches == nil {
-		err = fmt.Errorf("failed to parse 'parted' version:\n%w", err)
-		return 0, 0, err
-	}
-
-	major, _ := strconv.Atoi(matches[1])
-	minor, _ := strconv.Atoi(matches[2])
-
-	return major, minor, nil
-}
-
-// InitializeSinglePartition initializes a single partition based on the given partition configuration
-func InitializeSinglePartition(diskDevPath string, partitionNumber int,
-	partitionTableType configuration.PartitionTableType, partition configuration.Partition,
-) (partDevPath string, err error) {
-	const (
-		retryDuration    = time.Second
-		timeoutInSeconds = "5"
-		totalAttempts    = 5
-	)
-
-	partitionNumberStr := strconv.Itoa(partitionNumber)
-
-	// There are two primary partition naming conventions:
-	// - /dev/sdN<y>
-	// - /dev/loopNp<x>
-	// Detect the exact one we are using.
-	testPartDevPaths := []string{
-		fmt.Sprintf("%sp%s", diskDevPath, partitionNumberStr),
-	}
-
-	// If disk path ends in a digit, then the 'p<x>' style must be used.
-	// So, don't check the other style to avoid ambiguities. For example, /dev/loop1 vs. /dev/loop11.
-	// This is particularly relevant on Ubuntu, due to snap's use of loopback devices.
-	if !isDigit(diskDevPath[len(diskDevPath)-1]) {
-		devPath := fmt.Sprintf("%s%s", diskDevPath, partitionNumberStr)
-		testPartDevPaths = append(testPartDevPaths, devPath)
-	}
-
-	err = retry.Run(func() error {
-		for _, testPartDevPath := range testPartDevPaths {
-			exists, err := file.PathExists(testPartDevPath)
-			if err != nil {
-				err = fmt.Errorf("failed to find device path (%s):\n%w", testPartDevPath, err)
-				return err
-			}
-			if exists {
-				partDevPath = testPartDevPath
-				return nil
-			}
-			logger.Log.Debugf("Could not find partition path (%s). Checking other naming convention", testPartDevPath)
-		}
-		logger.Log.Warnf("Could not find any valid partition paths. Will retry up to %d times", totalAttempts)
-		err = fmt.Errorf("could not find partition to initialize in /dev")
-		return err
-	}, totalAttempts, retryDuration)
-
-	if err != nil {
-		return
-	}
-
-	logger.Log.Debugf("Initializing partition device path: %v", partDevPath)
-
-	// Set partition friendly name and partition type UUID (only for gpt)
-	if partitionTableType == configuration.PartitionTableTypeGpt {
-		setGptPartitionType(partition, timeoutInSeconds, diskDevPath, partitionNumberStr)
-	}
-
-	// Set partition flags if necessary
-	for _, flag := range partition.Flags {
-		args := []string{diskDevPath, "--script", "set", partitionNumberStr}
-		var flagToSet string
-		switch flag {
-		case configuration.PartitionFlagESP:
-			flagToSet = "esp"
-		case configuration.PartitionFlagGrub, configuration.PartitionFlagBiosGrub, configuration.PartitionFlagBiosGrubLegacy:
-			flagToSet = "bios_grub"
-		case configuration.PartitionFlagBoot:
-			flagToSet = "boot"
-		case configuration.PartitionFlagDeviceMapperRoot:
-			//Ignore, only used for internal tooling
-		default:
-			return partDevPath, fmt.Errorf("partition %v - Unknown partition flag: %v", partitionNumber, flag)
-		}
-		if flagToSet != "" {
-			args = append(args, flagToSet, "on")
-			// Golang does not allow mixing of variadic and regular arguments. So add all of the flock args to
-			// the overall arg slice and pass that to execute
-			args = append([]string{"--timeout", timeoutInSeconds, diskDevPath, "parted"}, args...)
-			_, stderr, err := shell.Execute("flock", args...)
-			if err != nil {
-				logger.Log.Warnf("Failed to set flag (%s) using parted: %v", flagToSet, stderr)
-			}
-		}
-	}
-
-	// Make sure all partition information is actually updated.
-	stdout, stderr, err := shell.Execute("flock", "--timeout", timeoutInSeconds, diskDevPath, "partprobe", "-s", diskDevPath)
-	if err != nil {
-		err = fmt.Errorf("failed to execute partprobe after partition initialization:\n%v\n%w", stderr, err)
-		return "", err
-	}
-	logger.Log.Debugf("Partprobe -s returned: %s", stdout)
-
-	return
-}
-
-func isDigit(c byte) bool {
-	return c >= '0' && c <= '9'
-}
-
-func setGptPartitionType(partition configuration.Partition, timeoutInSeconds, diskDevPath, partitionNumberStr string) (err error) {
-	if supports, _ := PartedSupportsTypeCommand(); !supports {
-		logger.Log.Warn("parted version <3.6 does not support the 'type' session command - skipping this operation")
-		return
-	}
-
-	if partition.TypeUUID != "" || partition.Type != "" {
-		var typeUUID string
-		if partition.TypeUUID != "" {
-			typeUUID = partition.TypeUUID
-		} else {
-			typeUUID = configuration.PartitionTypeNameToUUID[partition.Type]
-		}
-		_, stderr, err := shell.Execute("flock", "--timeout", timeoutInSeconds, diskDevPath, "parted", diskDevPath, "--script", "type", partitionNumberStr, typeUUID)
-		if err != nil {
-			logger.Log.Warnf("failed to set partition type using parted: %v", stderr)
-			err = nil
-			// Not-fatal
-		}
-	}
-	return
-}
-
-// FormatSinglePartition formats the given partition to the type specified in the partition configuration
-func FormatSinglePartition(partDevPath string, partition configuration.Partition,
-) (fsType string, err error) {
-	const (
-		totalAttempts = 5
-		retryDuration = time.Second
-	)
-
-	fsType = partition.FsType
-
-	// Note: It is possible for the format partition command to fail with error "The file does not exist and no size was specified".
-	// This is due to a possible race condition in Linux/parted where the partition may not actually be ready after being newly created.
-	// To handle such cases, we can retry the command.
-	switch fsType {
-	case "fat32", "fat16", "vfat", "ext2", "ext3", "ext4", "xfs":
-		mkfsOptions := DefaultMkfsOptions[fsType]
-
-		if fsType == "fat32" || fsType == "fat16" {
-			fsType = "vfat"
-		}
-
-		mkfsArgs := []string{"-t", fsType}
-		mkfsArgs = append(mkfsArgs, mkfsOptions...)
-		mkfsArgs = append(mkfsArgs, partDevPath)
-
-		err = retry.Run(func() error {
-			_, stderr, err := shell.Execute("mkfs", mkfsArgs...)
-			if err != nil {
-				logger.Log.Warnf("Failed to format partition using mkfs: %v", stderr)
-				return err
-			}
-
-			return err
-		}, totalAttempts, retryDuration)
-		if err != nil {
-			err = fmt.Errorf("could not format partition with type %v after %v retries", fsType, totalAttempts)
-		}
-	case "linux-swap":
-		err = retry.Run(func() error {
-			_, stderr, err := shell.Execute("mkswap", partDevPath)
-			if err != nil {
-				logger.Log.Warnf("Failed to format swap partition using mkswap: %v", stderr)
-				return err
-			}
-			return err
-		}, totalAttempts, retryDuration)
-		if err != nil {
-			err = fmt.Errorf("could not format partition with type %v after %v retries", fsType, totalAttempts)
-		}
-
-		_, stderr, err := shell.Execute("swapon", partDevPath)
-		if err != nil {
-			err = fmt.Errorf("failed to execute swapon:\n%v\n%w", stderr, err)
-			return "", err
-		}
-	case "":
-		logger.Log.Debugf("No filesystem type specified. Ignoring for partition: %v", partDevPath)
-	default:
-		return fsType, fmt.Errorf("unrecognized filesystem format: %v", fsType)
+		err = fmt.Errorf("unknown unit: %s", unit)
 	}
 
 	return
@@ -834,7 +163,7 @@ func SystemBlockDevices() (systemDevices []SystemBlockDevice, err error) {
 	includeFilter := strings.Join(blockDeviceMajorNumbers, ",")
 	rawDiskOutput, stderr, err := shell.Execute("lsblk", "-d", "--bytes", "-I", includeFilter, "-n", "--json", "--output", "NAME,SIZE,MODEL")
 	if err != nil {
-		err = fmt.Errorf("%v\n%w", stderr, err)
+		err = fmt.Errorf("failed to get disk list:\n%v\n%w", stderr, err)
 		return
 	}
 
@@ -842,12 +171,13 @@ func SystemBlockDevices() (systemDevices []SystemBlockDevice, err error) {
 	if rawDiskOutput != "" {
 		err = json.Unmarshal([]byte(rawDiskOutput), &blockDevices)
 		if err != nil {
+			err = fmt.Errorf("failed to parse disk list:\n%w", err)
 			return
 		}
 	}
 
 	if len(blockDevices.Devices) <= 0 {
-		err = fmt.Errorf("failed to find supported disks:\n%w", err)
+		err = fmt.Errorf("no disks found")
 		return
 	}
 
@@ -855,166 +185,15 @@ func SystemBlockDevices() (systemDevices []SystemBlockDevice, err error) {
 
 	for i, disk := range blockDevices.Devices {
 		systemDevices[i].DevicePath = fmt.Sprintf("/dev/%s", disk.Name)
-
-		systemDevices[i].RawDiskSize, err = strconv.ParseUint(disk.Size.String(), 10, 64)
-		if err != nil {
-			return
-		}
-
 		systemDevices[i].Model = strings.TrimSpace(disk.Model)
-	}
 
-	return
-}
-
-func GetDiskPartitions(diskDevPath string) ([]PartitionInfo, error) {
-	// Just in case the disk was only recently connected, wait for the OS to finish processing it.
-	err := WaitForDevicesToSettle()
-	if err != nil {
-		return nil, fmt.Errorf("failed to list disk (%s) partitions:\n%w", diskDevPath, err)
-	}
-
-	// Read the disk's partitions.
-	jsonString, _, err := shell.Execute("lsblk", diskDevPath, "--output", "NAME,PATH,PARTTYPE,FSTYPE,UUID,MOUNTPOINT,PARTUUID,PARTLABEL,TYPE", "--json", "--list")
-	if err != nil {
-		return nil, fmt.Errorf("failed to list disk (%s) partitions:\n%w", diskDevPath, err)
-	}
-
-	var output partitionInfoOutput
-	if jsonString != "" {
-		err = json.Unmarshal([]byte(jsonString), &output)
+		// Parse the size (which could be a quoted string or JSON number)
+		size, err := disk.Size.Int64()
 		if err != nil {
-			return nil, fmt.Errorf("failed to parse disk (%s) partitions JSON:\n%w", diskDevPath, err)
+			err = fmt.Errorf("failed to parse disk size for %s: %w", disk.Name, err)
+			return systemDevices, err
 		}
-	}
-
-	return output.Devices, err
-}
-
-func createExtendedPartition(diskDevPath string, partitionTableType configuration.PartitionTableType,
-	partitions []configuration.Partition, partIDToFsTypeMap, partDevPathMap map[string]string,
-	partedSupportsEmptyStringArgs bool,
-) (err error) {
-	// Create a new partition object for extended partition
-	extendedPartition := configuration.Partition{}
-	extendedPartition.ID = extendedPartitionType
-	extendedPartition.Start = partitions[maxPrimaryPartitionsForMBR-1].Start
-	extendedPartition.End = partitions[len(partitions)-1].End
-
-	partDevPath, err := createSinglePartition(diskDevPath, maxPrimaryPartitionsForMBR, partitionTableType,
-		extendedPartition, extendedPartitionType, partedSupportsEmptyStringArgs)
-	if err != nil {
-		err = fmt.Errorf("failed to create extended partition:\n%w", err)
-		return
-	}
-	partIDToFsTypeMap[extendedPartition.ID] = ""
-	partDevPathMap[extendedPartition.ID] = partDevPath
-	return
-}
-
-func getPartUUID(device string) (uuid string, err error) {
-	stdout, _, err := shell.Execute("blkid", device, "-s", "UUID", "-o", "value")
-	if err != nil {
-		return
-	}
-	logger.Log.Trace(stdout)
-	uuid = strings.TrimSpace(stdout)
-	return
-}
-
-func getSectorSizeFromFile(sectorFile string) (sectorSize uint64, err error) {
-	if exists, ferr := file.PathExists(sectorFile); ferr != nil {
-		err = fmt.Errorf("failed to access sector size file (%s):\n%w", sectorFile, ferr)
-		return
-	} else if !exists {
-		err = fmt.Errorf("could not find the hw sector size file %s to obtain the sector size of the system", sectorFile)
-		return
-	}
-
-	fileContent, err := file.ReadLines(sectorFile)
-	if err != nil {
-		err = fmt.Errorf("failed to read from (%s):\n%w", sectorFile, err)
-		return
-	}
-
-	// sector file should only have one line, return error if not
-	if len(fileContent) != 1 {
-		err = fmt.Errorf("%s has more than one line", sectorFile)
-		return
-	}
-
-	sectorSize, err = strconv.ParseUint(fileContent[0], 10, 64)
-	return
-}
-
-func GetSectorSize(diskDevPath string) (logicalSectorSize, physicalSectorSize uint64, err error) {
-	const (
-		diskNameStartIndex = 5
-	)
-
-	// Grab the specific disk name from /dev/xxx
-	matchResult := diskDevPathRegexp.MatchString(diskDevPath)
-	if !matchResult {
-		err = fmt.Errorf("input disk device path (%s) is of invalud format", diskDevPath)
-		return
-	}
-	diskName := diskDevPath[diskNameStartIndex:len(diskDevPath)]
-
-	hw_sector_size_file := fmt.Sprintf("/sys/block/%s/queue/hw_sector_size", diskName)
-	physical_sector_size_file := fmt.Sprintf("/sys/block/%s/queue/physical_block_size", diskName)
-
-	logicalSectorSize, err = getSectorSizeFromFile(hw_sector_size_file)
-	if err != nil {
-		return
-	}
-
-	physicalSectorSize, err = getSectorSizeFromFile(physical_sector_size_file)
-	return
-}
-
-func alignSectorAddress(sectorAddr, logicalSectorSize, physicalSectorSize uint64) (alignedSector uint64) {
-	// Need to make sure that starting sector of a partition is aligned based on the physical sector size of the system.
-	// For example, suppose the physical sector size is 4096. If the input start sector is 40960001, then this is misaligned,
-	// and need to be elevated to the next aligned address, which is (40960001/4096 + 1)*4096 = 4100096.
-
-	// We do need to take care of a special case, which is the first partition (normally boot partition) might be less than
-	// the physical sector size. In this case, we need to check whether the start sector is a multiple of 1 MiB.
-	alignedSector = 0
-	if sectorAddr < physicalSectorSize {
-		if sectorAddr%(MiB/logicalSectorSize) == 0 {
-			alignedSector = sectorAddr
-		}
-	} else if (sectorAddr % physicalSectorSize) == 0 {
-		alignedSector = sectorAddr
-	} else {
-		alignedSector = (sectorAddr/physicalSectorSize + 1) * physicalSectorSize
-	}
-
-	return
-}
-
-func obtainPartitionDetail(partitionIndex int, hasExtendedPartition bool) (partType string, partitionNumber int) {
-	const (
-		indexOffsetForNormalPartitionNumber  = 1
-		indexOffsetForLogicalPartitionNumber = 2
-	)
-
-	// partitionIndex is the index of the partition in the partition array, which starts at 0.
-	// partitionNumber, however, starts at 1 (E.g. /dev/sda1), and thus partitionNumber = partitionIndex + 1.
-	// In the case of logical partitions, since an extra extended partition has to be created first in order to
-	// to create logical partitions, so the partition number will further increase by 1, which equals partitionIndex + 2.
-
-	if hasExtendedPartition && partitionIndex >= (maxPrimaryPartitionsForMBR-1) {
-		if partitionIndex == (maxPrimaryPartitionsForMBR - 1) {
-			partType = extendedPartitionType
-			partitionNumber = partitionIndex + indexOffsetForNormalPartitionNumber
-		} else {
-			partType = logicalPartitionType
-			partitionNumber = partitionIndex + indexOffsetForLogicalPartitionNumber
-		}
-	} else {
-		partType = primaryPartitionType
-		partitionNumber = partitionIndex + indexOffsetForNormalPartitionNumber
+		systemDevices[i].RawDiskSize = uint64(size)
 	}
 
 	return
